Extracted OpinionChart as component
